3-connector-rest-apis/gcp_api/requirements.txt
# Function dependencies, for example:
# package>=version
Flask==2.0.2
requests
google-api-core==2.4.0
google-api-python-client==1.7.9
google-auth==2.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-storage==2.1.0